Fixed errors from refactor
